Update ARM64 / Raspbian status to broken
Updated ARM64 / Raspbian status to BROKEN
Reason:  The method to run the server uses emulation.  This emulation requires a set of binaries that are relatively close to the ones that were used to build the current version of the server.  The prepackaged ones included in this repository are a couple of years old and probably aren't even the right GCC version anymore.

2 years ago this was too slow to be worth it.  You could not even join the server before timing out most of the time and when you could join the map never really loaded etc.  Due to the releases of several new Pi models with a lot more memory and power as well as lots of advancements in QEMU and other emulation software it may be worth trying again later, but for now I am marking these platforms as broken.