Reimplement custom directory support
Only works with directories that the user has access to. Lets say the user can create a folder with mkdir in, for example, ~/_data/MyData/. If the user wants the Minecraft Server to be saved in the MyData folder and the user has access to create folders there without sudo, the script will be able to do so too.

If the user wants to have his Minecraft Server on ~/_data/sudoexample/, and he cannot create any additional folder in sudoexample without sudo, then the script will also be unable to.